+++ src/models/ftl-req.ts
@@ -1,0 +1,5 @@
+import {FtlValue} from './ftl-value';
+
+export class FtlReq extends FtlValue {
+  readonly kind = 'Custom Requirement';
+}